requirements.txt
gradio
onnxruntime
onnx
torch
torchvision
pillow
numpy